+++ lib/openapi_first/runtime_response.rb
@@ -11,10 +11,18 @@ module OpenapiFirst
     def initialize(operation, rack_response)
       @operation = operation
       @rack_response = rack_response
+      @error = nil
     end
+
+    attr_reader :error
 
     def_delegators :@rack_response, :status, :content_type
     def_delegators :@operation, :name
+
+    def valid?
+      validate unless @validated
+      @error.nil?
+    end
 
     def known?
       !!response_definition
@@ -37,7 +45,8 @@ module OpenapiFirst
     end
 
     def validate
-      ResponseValidation::Validator.new(@operation).validate(self)
+      @validated = true
+      @error = ResponseValidation::Validator.new(@operation).validate(self)
     end
 
     def validate!